+++ src/main/java/prgms/vouchermanagementapp/controller/CommandExecutor.java
@@ -3,11 +3,11 @@ package prgms.vouchermanagementapp.controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
-import prgms.vouchermanagementapp.domain.VoucherType;
-import prgms.vouchermanagementapp.domain.value.Amount;
-import prgms.vouchermanagementapp.domain.value.Ratio;
 import prgms.vouchermanagementapp.exception.IllegalCommandException;
 import prgms.vouchermanagementapp.exception.IllegalVoucherTypeIndexException;
+import prgms.vouchermanagementapp.model.VoucherType;
+import prgms.vouchermanagementapp.model.value.Amount;
+import prgms.vouchermanagementapp.model.value.Ratio;
 import prgms.vouchermanagementapp.service.VoucherService;
 import prgms.vouchermanagementapp.view.CommandType;
 import prgms.vouchermanagementapp.view.IoManager;
