+++ src/main/java/prgms/vouchermanagementapp/model/value/Amount.java
@@ -1,4 +1,4 @@
-package prgms.vouchermanagementapp.domain.value;
+package prgms.vouchermanagementapp.model.value;
 
 import java.text.MessageFormat;
 import java.util.Objects;
